Workaround for upgrades from OI before 2017.10 to the latest code
OpenIndiana releases before 2017.10 have tools which are unable
to cope with installing later kernels. Basically, tools before fix
for 8142 was integrated are unable to install kernel which has fix
for 8685 integraded.

So, OpenIndiana 2017.04 (and older) can't be updated directly to
OpenIndiana 2018.04 (and later releases). You should update thru
OpenIndiana 2017.10, or use the workaround below.

For the discussion in full, see:
  * https://www.illumos.org/issues/8142
  * https://www.illumos.org/issues/8685
  * https://illumos.topicbox.com/groups/developer/T441d169a85277364-Md5eecf7b8e4e78766030cfe2
  * https://illumos.topicbox.com/groups/developer/Tbd6485b901b93374-M2c975c580ad03866bfaf188e